+++ adalflow/adalflow/utils/lazy_import.py
@@ -71,6 +71,7 @@ class OptionalPackages(Enum):
         ],
         "Please install Azure packages with: pip install azure-identity azure-core azure-ai-formrecognizer azure-ai-textanalytics",
     )
+    TOGETHER = ("together", "Please install together with: pip install together")
     # search library
     FAISS = (
         "faiss",